added dry-run mode.

+++ rlayout/exporters/lfs_directory.rb
@@ -13,13 +13,14 @@ module RLayout
       
       # Initialize with directory path to generate to
       def initialize(directory_path, opts = {})
-        @opts = {:delete_if_existing => false, :chunksize => 1024 }.merge(opts)
+        @opts = {:dry_run => false, :delete_if_existing => false, :chunksize => 1024 }.merge(opts)
         @dir_path = File.expand_path(directory_path)        
         @chunksize = @opts[:chunksize]
       end
       
       def prologue(root)
-        FileUtils.mkdir_p(@dir_path)
+        
+        fu_mode.mkdir_p(@dir_path)
         @clean_up_path = File.join(@dir_path, root.name)
         cleanup if @opts[:delete_if_existing]
         @dir_path
@@ -34,7 +35,7 @@ module RLayout
       end
       
       def cleanup
-         FileUtils.rm_rf(@clean_up_path) if File.exist?(@clean_up_path)
+        fu_mode::rm_rf(@clean_up_path) if File.exist?(@clean_up_path)
       end
      
       protected
@@ -42,15 +43,29 @@ module RLayout
       # Process a single node
       def process_node(path, node)
         if node.kind_of?(VFSGroup)
-          FileUtils.mkdir_p(path)
+          fu_mode.mkdir_p(path)
         else
-          File.open(path, "wb") do |ios|
-            node.read_stream(@chunksize) { |bytes| ios.write(bytes) }
+          if (!@opts[:dry_run])
+            File.open(path, "wb") do |ios|
+              node.read_stream(@chunksize) { |bytes| ios.write(bytes) }
+            end
+          else
+            puts "write #{path}"
           end
         end
       end
+      
+      # Fileutils mode.
+      def fu_mode
+        if (@opts[:dry_run])
+          FileUtils::DryRun
+        else
+          FileUtils
+        end
+      end
+      
     end
-    
+  
     def Exporters.lfs_directory(directory_path, opts = {})
       LFSDirectory.new(directory_path, opts)
     end
